requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
httpx==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
transformers>=4.33.2
torch>=2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
